Major refactor for table display, corrected binding for EGT data, removed Person class (placeholder for early development)
Changed table display to account for Chrome/Edge browsers not showing sticky headers correctly:

Now displaying two tables, one with null data rows, the other with a hidden header row.

In order to make this work, horizontal scroll events for both tables needed to be bound to one another (in the html).

This doesn't seem to be optimal for Firefox, but users will most likely be on Chrome/Edge anyway.